app/requirements.txt
google-cloud-storage
pillow
sendgrid
tensorflow-hub
